fix: editing with empty end
